% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{a}
\alias{a}
\title{Sequence data for the dissected and sectioned gut regions of Drosophila melanogaster}
\format{
A data frame  with 17559 rows and 2 variables:
\describe{
\item{GeneID}{Flybase gene ID}
\item{readCount}{Raw read counts}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
a
}
\description{
Sequence data the SRR891601 run of the subregions a1-3
}
\keyword{datasets}
