% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{a1_vs_p1_clusters}
\alias{a1_vs_p1_clusters}
\title{ClustrProfiler data for the midgut data}
\format{
A ClustrProfiler object containing results of KEGG ontology based over-representation test:
\describe{
\item{organism}{dme - Drosophila melanogaster}
\item{ontology}{KEGG}
\item{keytype}{ncbi-geneid}
\item{gene}{A list of NCBI gene IDs}
\item{ID}{List of KEGG Drosophila melanogaster pathway IDs}
\item{Description}{List of names of associated KEGG pathways}
\item{Gene Ratio}{Ratio of input genes that are annotated in the associated term. Count of significant genes over total number in the category}
\item{BgRatio}{Ratio of all genes that are annotated in the associated term}
\item{pvalue}{p-value for test that determines the given biological functions or processes are over-represented or enriched}
\item{p.adjust}{The adjusted p-value. FDR is used by default}
\item{qvalue}{An alternative to FDR control for multiple testing}
\item{geneID}{List of NIH gene IDs}
\item{Count}{Count of significant genes in a given gene set}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
a1_vs_p1_clusters
}
\description{
enrichResult object for midgut data cached Dec 1 2021
}
\keyword{datasets}
