% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{i}
\alias{i}
\title{This data set contains unweighted PUMS census data from the Los Angeles and Long Beach areas for the years 1970, 1980, and 1990}
\format{
A data frame  with 6 rows and 61 variables:
\describe{
\item{X1}{YEAR}
\item{X2}{GQ}
\item{X3}{GQTYPE}
\item{X4}{FARM}
\item{X5}{OWNERSHP}
\item{X6}{VALUE}
\item{X7}{RENT}
\item{X8}{FTOTINC}
\item{X9}{NFAMS}
\item{X10}{NCOUPLES}
\item{X11}{NMOTHERS}
\item{X12}{NFATHERS}
\item{X13}{MOMLOC}
\item{X14}{STEPMOM}
\item{X15}{MOMRULE}
\item{X16}{POPLOC}
\item{X17}{STEPPOP}
\item{X18}{POPRULE}
\item{X19}{SPLOC}
\item{X20}{SPRULE}
\item{X21}{FAMSIZE}
\item{X22}{NCHILD}
\item{X23}{NCHLT5}
\item{X24}{FAMUNIT}
\item{X25}{ELDCH}
\item{X26}{YNGCH}
\item{X27}{NSIBS}
\item{X28}{RELATE}
\item{X29}{AGE}
\item{X30}{SEX}
\item{X31}{RACE}
\item{X32}{MARST}
\item{X33}{CHBORN}
\item{X34}{BPL}
\item{X35}{SCHOOL}
\item{X36}{EDUCREC}
\item{X37}{SCHLTYPE}
\item{X38}{EMPSTAT}
\item{X39}{LABFORCE}
\item{X40}{OCC1950}
\item{X41}{OCCSCORE}
\item{X42}{SEI}
\item{X43}{IND1950}
\item{X44}{CLASSWKR}
\item{X45}{WKSWORK2}
\item{X46}{HRSWORK2}
\item{X47}{YRLASTWK}
\item{X48}{WORKEDYR}
\item{X49}{INCTOT}
\item{X50}{INCWAGE}
\item{X51}{INCBUS}
\item{X52}{INCFARM}
\item{X53}{INCSS}
\item{X54}{INCWELFR}
\item{X55}{INCOTHER}
\item{X56}{POVERTY}
\item{X57}{MIGRATE5}
\item{X58}{MIGPLAC5}
\item{X59}{MOVEDIN}
\item{X60}{VETSTAT}
\item{X61}{TRANWORK}
}
}
\source{
\url{http://kdd.ics.uci.edu/databases/ipums/ipums.html}
}
\usage{
i
}
\description{
Unweughted PUMS for the year 1990
}
\keyword{datasets}
