% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData-package.R
\docType{package}
\name{MarianesMidgutData-package}
\alias{MarianesMidgutData}
\alias{MarianesMidgutData-package}
\title{MarianesMidgutData: Data Package for the sequence date of the Drosophila Midgut expression from Marianes et al. 2013}
\description{
Loads the R objects for data associated with "Physiological and stem cell compartmentalization within the Drosophila midgut" by Marianes & Spradling (2013) (https://elifesciences.org/articles/00886)"
}
\author{
\strong{Maintainer}: Javier Carpinteyro-Ponce \email{jcarpinteyro@carnegiescience.edu} (\href{https://orcid.org/0009-0005-2865-1475}{ORCID})

Authors:
\itemize{
  \item Katie Cox \email{katherinecox@jhu.edu}
  \item Fred Tan \email{tan@carnegiescience.edu}
  \item Sayumi York \email{syork@ndm.edu}
}

}
\keyword{internal}
