% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{midgut}
\alias{midgut}
\title{Expression data for the midgut data}
\format{
A DESeq2 object containing results of differential expression analysis:
\describe{
\item{assays}{List of matrices with raw and normalized count data}
\item{rowRanges}{Genomic ranges for each gene}
\item{colData}{Data frame with sample metadata}
\item{design}{Formula representing the experimental design}
\item{dispersions}{Estimated dispersion values}
\item{results}{Data frame with differential expression results}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
midgut
}
\description{
DESeq2 object for the midgut data
}
\keyword{datasets}
