% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{midgut_tsv}
\alias{midgut_tsv}
\title{Sample meta data for the midgut data}
\format{
A data frame  with 30 rows and 3 variables:
\describe{
\item{sample}{Sample name}
\item{filename}{File name}
\item{condition}{Condition}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
midgut_tsv
}
\description{
Sample meta data for the midgut data
}
\keyword{datasets}
