% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{SRR891602}
\alias{SRR891602}
\title{Sequence data for the dissected and sectioned gut regions of Drosophila melanogaster}
\format{
A data frame  with 17559 rows and 2 variables:
\describe{
\item{V1}{Flybase gene ID}
\item{V2}{Raw read counts}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
SRR891602
}
\description{
Sequence data the SRR891601 run of the subregions a1-3 replicate 2
}
\keyword{datasets}
