% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarianesMidgutData.R
\docType{data}
\name{a1_vs_p1_clusters}
\alias{a1_vs_p1_clusters}
\title{Sample meta data for the midgut data}
\format{
A DOSE object containing results of differential expression analysis:
\describe{
\item{sample}{Sample name}
}
}
\source{
\url{https://elifesciences.org/articles/00886}
}
\usage{
a1_vs_p1_clusters
}
\description{
Sample meta data for the midgut data
}
\keyword{datasets}
